requirements.txt
flask
flask-sqlalchemy
anthropic
slack-bolt
yt-dlp
python-dotenv
requests
openai-whisper
